improved naming of base config, separated processor loader code, finalized testing of recall booster option
