+++ base_config.json
@@ -10,7 +10,7 @@
       false
     ]
   },
-  "recall_boost": false,
+  "use_recall_boost": false,
   "redactor_open_char": "[",
   "redactor_close_char": "]",
   "annotators": {
@@ -154,10 +154,11 @@
         "tag": "voornaam",
         "overlapping": true,
         "lookup_values": "first_name",
-        "recall_boost": {
-          "expander": "docdeid.str.expander.MinimumLengthExpander",
+        "recall_boost_config": {
+          "recall_boost_type": "docdeid.str.expander.MinimumLengthExpander",
           "args":{
-            "min_len": 4
+            "min_len": 4,
+            "str_processors": ["deduce.str.processor.TitleCase"]
           }
         }
       }
@@ -169,8 +170,12 @@
         "tag": "achternaam",
         "overlapping": true,
         "lookup_values": "surname",
-        "recall_boost": {
-          "recall_boost_type": "docdeid.str.expander.MinimumLengthExpander"  
+        "recall_boost_config": {
+          "recall_boost_type": "docdeid.str.expander.MinimumLengthExpander",
+          "args": {
+            "min_len": 4,
+            "str_processors": ["deduce.str.processor.TitleCase"]
+          }  
         }
       }
     },
